+++ src/main/java/vicky/command/MarkTaskCommand.java
@@ -1,29 +1,30 @@
-package command;
+package vicky.command;
 
-import TaskList.TaskList;
-import TaskList.Task;
-import storage.Storage;
-import ui.Ui;
+import vicky.taskList.TaskList;
+import vicky.taskList.Task;
+import vicky.storage.Storage;
+import vicky.ui.Ui;
 
 import java.io.IOException;
 
-public class UnmarkTaskCommand extends Command {
+public class MarkTaskCommand extends Command {
 
     private int index;
 
-    public UnmarkTaskCommand(int index) {
+    public MarkTaskCommand(int index) {
         this.index = index;
     }
 
     @Override
     public String execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
         try {
-            Task t = tasks.unmarkTask1(this.index);
+            Task t = tasks.markTask1(this.index);
             storage.save(tasks);
-            return ui.showUnmarkedTask(t);
+            return ui.showMarkedTask(t);
         } catch (IndexOutOfBoundsException e) {
             return ui.showIndexOutOfBounds();
         }
+
     }
 
     @Override
