+++ src/main/java/vicky/command/UnmarkTaskCommand.java
@@ -1,30 +1,29 @@
-package command;
+package vicky.command;
 
-import TaskList.TaskList;
-import TaskList.Task;
-import storage.Storage;
-import ui.Ui;
+import vicky.taskList.TaskList;
+import vicky.taskList.Task;
+import vicky.storage.Storage;
+import vicky.ui.Ui;
 
 import java.io.IOException;
 
-public class MarkTaskCommand extends Command {
+public class UnmarkTaskCommand extends Command {
 
     private int index;
 
-    public MarkTaskCommand(int index) {
+    public UnmarkTaskCommand(int index) {
         this.index = index;
     }
 
     @Override
     public String execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
         try {
-            Task t = tasks.markTask1(this.index);
+            Task t = tasks.unmarkTask1(this.index);
             storage.save(tasks);
-            return ui.showMarkedTask(t);
+            return ui.showUnmarkedTask(t);
         } catch (IndexOutOfBoundsException e) {
             return ui.showIndexOutOfBounds();
         }
-
     }
 
     @Override
