+++ src/main/java/vicky/exception/InvalidTaskException.java
@@ -1,4 +1,4 @@
-package TaskList;
+package vicky.exception;
 
 public class InvalidTaskException extends RuntimeException {
     public InvalidTaskException(String message) {
