+++ src/main/java/vicky/taskList/Deadline.java
@@ -1,6 +1,5 @@
-package TaskList;
+package vicky.taskList;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
