+++ src/main/java/vicky/taskList/Event.java
@@ -1,6 +1,5 @@
-package TaskList;
+package vicky.taskList;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 public class Event extends Deadline {
