+++ src/main/java/vicky/taskList/Todo.java
@@ -1,4 +1,4 @@
-package TaskList;
+package vicky.taskList;
 
 public class Todo extends Task {
